feat: Connection status propagation

+++ app/src/main/java/me/worric/souvenarius/data/repository/StorageHandler.java
@@ -3,6 +3,8 @@ package me.worric.souvenarius.data.repository;
 import android.net.Uri;
 import android.support.annotation.NonNull;
 
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 
@@ -16,16 +18,23 @@ import timber.log.Timber;
 public class StorageHandler {
 
     private static final String STORAGE_REFERENCE = "images";
+    private final FirebaseAuth mAuth;
     private final FirebaseStorage mFirebaseStorage;
     private final StorageReference mRef;
 
     @Inject
     public StorageHandler() {
+        mAuth = FirebaseAuth.getInstance();
         mFirebaseStorage = FirebaseStorage.getInstance();
         mRef = mFirebaseStorage.getReference(STORAGE_REFERENCE);
     }
 
     public void uploadImage(@NonNull File imageFile) {
+        if (checkAuthState() == null) {
+            Timber.w("User is null, not uploading image");
+            return;
+        }
+
         StorageReference ref = mRef.child(imageFile.getName());
         ref.putFile(Uri.fromFile(imageFile)).addOnCompleteListener(task -> {
             if (task.isSuccessful()) {
@@ -42,6 +51,11 @@ public class StorageHandler {
 
     public void removeImage(@NonNull String photoName) {
         Timber.i("attempting deletion of A SINGLE photos of the souvenir");
+        if (checkAuthState() == null) {
+            Timber.w("User is null, not removing image");
+            return;
+        }
+
         mRef.child(photoName).delete().addOnCompleteListener(task -> {
             Timber.i("The DELETE task is successful? %s", task.isSuccessful());
             if (!task.isSuccessful()) Timber.e(task.getException(), "Task DELETE didn't execute right!");
@@ -49,6 +63,11 @@ public class StorageHandler {
     }
 
     public void removeImages(@NonNull List<String> photos) {
+        if (checkAuthState() == null) {
+            Timber.w("User is null, not removing multiple images");
+            return;
+        }
+
         if (photos.isEmpty()) {
             Timber.w("list of photos for the current souvenir is empty. Skipping deletion...");
             return;
@@ -62,4 +81,9 @@ public class StorageHandler {
             });
         }
     }
+
+    private FirebaseUser checkAuthState() {
+        return mAuth.getCurrentUser();
+    }
+
 }
